Fix all clippy warnings and errors
- Remove unused imports across all modules
- Fix unused variables (prefix with underscore)
- Fix unused mut warnings
- Fix type_complexity warnings (add allow attributes)
- Fix dead_code warnings for future-use items (allow attributes)
- Fix redundant pattern matching (use is_err())
- Fix field assignment outside initializer (use struct init)
- Fix useless vec! macro in example
- Rename from_str to parse to avoid confusion with FromStr trait
- Remove needless fn main in doctest
- Add StringCallback type alias for widget callbacks
- Fix non-exhaustive pattern match in layout

+++ src/render/text.rs
@@ -84,8 +84,8 @@ impl TextRenderer {
                 let physical_glyph = glyph.physical((0.0, 0.0), 1.0);
 
                 if let Some(image) = self.swash_cache.get_image(&mut font_system, physical_glyph.cache_key) {
-                    let glyph_x = physical_glyph.x as i32;
-                    let glyph_y = physical_glyph.y as i32 + run.line_y as i32;
+                    let glyph_x = physical_glyph.x;
+                    let glyph_y = physical_glyph.y + run.line_y as i32;
 
                     for (i, alpha) in image.data.iter().enumerate() {
                         let px = glyph_x + (i as i32 % image.placement.width as i32);
